modify delete from db

+++ src/main/java/com/emmkay/infertility_system_api/service/AdminService.java
@@ -148,6 +148,9 @@ public class AdminService {
     public void deleteUserFromDatabase(String userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTED));
+        if (!user.getIsRemoved()) {
+            throw new AppException(ErrorCode.USER_ALREADY_ACTIVE);
+        }
         userRepository.delete(user);
     }
 }
